Moved /helper out of /pages to get next.js build to work
