+++ helper/weatherDataHelper.ts
